+++ oap-server/analyzer/agent-analyzer/src/main/java/org/apache/skywalking/oap/server/analyzer/provider/meter/config/MeterConfig.java
@@ -21,6 +21,7 @@ package org.apache.skywalking.oap.server.analyzer.provider.meter.config;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.apache.skywalking.oap.meter.analyzer.MetricRuleConfig;
+import org.apache.skywalking.oap.meter.analyzer.k8s.K8sInfoRegistry;
 
 import java.util.List;
 
@@ -32,6 +33,13 @@ public class MeterConfig implements MetricRuleConfig {
     private String expPrefix;
     private String filter;
     private List<Rule> metricsRules;
+    private boolean kubernetesAware;
+
+    public void init() {
+        if (this.kubernetesAware) {
+            K8sInfoRegistry.getInstance().start();
+        }
+    }
 
     @Data
     @NoArgsConstructor